copy_engine: rename spdk_copy_module_get_max_ctx_size() to spdk_copy_task_size()
Make the public API clearer - if the user wants to allocate a
spdk_copy_task directly, they need to allocate spdk_copy_task_size()
bytes.

Also change the return type to size_t for consistency.

+++ include/spdk_internal/copy_engine.h
@@ -74,7 +74,7 @@ struct spdk_copy_module_if {
 	 */
 	void	(*config_text)(FILE *fp);
 
-	int	(*get_ctx_size)(void);
+	size_t	(*get_ctx_size)(void);
 	TAILQ_ENTRY(spdk_copy_module_if)	tailq;
 };
